Fix warnings related to GPRS debugging messages.
git-svn-id: http://voip.null.ro/svn/yatebts@474 26ef948f-17ed-0410-91a8-c94bd273f071

+++ mbts/GPRS/MAC.cpp
@@ -655,10 +655,10 @@ static void dumpPdch()
 {
 
 	PDCHL1FEC *ch;
-	printf("PDCHs=%d:",gL2MAC.macPDCHs.size());
+	printf("PDCHs=%u:",(unsigned int)gL2MAC.macPDCHs.size());
 	RN_MAC_FOR_ALL_PDCH(ch) { printf(" %s",ch->shortId()); }
 	printf("\n");
-	printf("PACCHs=%d",gL2MAC.macPacchs.size());
+	printf("PACCHs=%u",(unsigned int)gL2MAC.macPacchs.size());
 	RN_MAC_FOR_ALL_PACCH(ch) { printf(" %s",ch->shortId()); }
 	printf("\n");
 }
