+++ mbts/GPRS/RLCEngine.cpp
@@ -1067,7 +1067,7 @@ RLCDownlinkDataBlock* RLCDownEngine::engineFillBlock(unsigned bsn,
 		// Add the pdu segments.
 		for (int j = 0; j < pducnt; j++) {
 			block->mPayload.append(pdus[j]);
-			if (GPRSDebug) sprintf(report+strlen(report)," seg=%d",pdus[j].size());
+			if (GPRSDebug) sprintf(report+strlen(report)," seg=%u",(unsigned int)pdus[j].size());
 		}
 		// Add filler, if any.  Unused RLC data field filled with 0x2b as per 04.60 10.4.16
 		int fillsize = payloadsize - block->mPayload.size();
